fix: move active-runner tests to tests/, fix defaultModelTier undefined coalescing (WOP-1899)
